server/requirements.txt
fastapi>=0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic>=2.7.4
python-jose==3.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pyjwt==2.6.0
python-multipart==0.0.6

httpx>=0.23.0,<1
starlette>=0.27.0
openai==1.73.0
pydantic[email]
shortuuid

# Updated to newer compatible versions
langchain>=0.3.23
langchain-openai>=0.3.13
langchain-core>=0.3.52
typing-extensions>=4.8.0
tenacity>=8.2.0
langsmith>=0.3.31
langchain-community>=0.3.21
langchain-google-community>=2.0.7
langgraph>=0.3.30
snowflake